requirements.txt
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
deepspeed==0.15.3
dill==0.3.8
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.5.8/flash_attn-2.5.8+cu122torch2.3cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=13454dd3d37cf173649bd389b84614b8072fb283f8f2fd23a65ab66caafc304b
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.6.1
gradio==4.42.0
gradio_client==1.3.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
-e git+https://github.com/hiyouga/LLaMA-Factory@55691251c2da659d69e10fa1ce32807cf80bf951#egg=llamafactory
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
orjson==3.10.10
packaging==24.1
pandas==2.2.3
peft==0.12.0
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
qwen-vl-utils==0.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
ruff==0.7.1
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.41.2
sympy==1.13.3
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.6
transformers @ git+https://github.com/huggingface/transformers@21fac7abba2a37fae86106f87fcf9974fd1e3830
transformers-stream-generator==0.0.4
triton==3.0.0
trl==0.9.6
typer==0.12.5
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
xxhash==3.5.0
yarl==1.17.0
seaborn